+++ docs/content/News.rst
@@ -8,6 +8,7 @@ snakePipes 2.x.y
 * Added filtering of empty drops with Dropletutils to scRNA-seq mode STARsolo
 * Added spikein normalization to ChIPseq workflow
 * Added hybrid genome creation to createIndices
+* Added STARsolo report for all samples to STARsolo output folder
 * FASTQ1 and FASTQ2 are not localrules anymore due to buggy logging
 * Included optional differential splicing analysis using rmats within mRNA-seq workflow
 * Symlinks in the output path are relative